Allow optional output properties
